+++ app/src/components/composable/convertCoordinates.js
@@ -97,3 +97,41 @@ export function create_multibbox(contours) {
 
   return { minX, minY, maxX, maxY }
 }
+
+export function getDynamicTolerance(polygon) {
+  const len = polygon.length
+
+  if (len > 50 && len <= 100) {
+    return 0.3
+  } else if (len > 20 && len <= 50) {
+    return 0.1
+  } else if (len > 10 && len <= 20) {
+    return 0.05
+  } else if (len > 100 && len <= 1000) {
+    return 1
+  } else if (len > 1000 && len <= 2000) {
+    return 5
+  } else if (len > 2000 && len <= 5000) {
+    return 10
+  } else if (len > 5000) {
+    return 20
+  }
+
+  return len > 20 ? 2 : 0.001;
+}
+
+export function roundCoordinates(multiPolygon, precision = 6) {
+  return multiPolygon.getCoordinates().map(polygon =>
+    polygon.map(ring =>
+      ring.map(coord => {
+        const longitude = parseFloat(coord[0]);
+        const latitude = parseFloat(coord[1]);
+
+        return [
+          longitude.toFixed(precision),
+          latitude.toFixed(precision)
+        ];
+      })
+    )
+  );
+}
